+++ elasticsearch7/search-index/src/main/java/org/vertexium/elasticsearch7/utils/InfiniteScrollIterable.java
@@ -9,12 +9,19 @@ import org.vertexium.query.IterableWithScores;
 import org.vertexium.query.QueryResultsIterable;
 import org.vertexium.util.CloseableIterator;
 import org.vertexium.util.CloseableUtils;
+import org.vertexium.util.VertexiumLogger;
+import org.vertexium.util.VertexiumLoggerFactory;
 
 import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.Iterator;
 import java.util.NoSuchElementException;
+import java.util.stream.Collectors;
 
 public abstract class InfiniteScrollIterable<T> implements QueryResultsIterable<T>, IterableWithScores<T> {
+    private static final VertexiumLogger LOGGER = VertexiumLoggerFactory.getLogger(InfiniteScrollIterable.class);
+    private static final String SCROLL_API_STACK_TRACE_LOGGER_NAME = "org.vertexium.elasticsearch7.SCROLL_API_STACK_TRACE";
+    private static final VertexiumLogger SCROLL_API_STACK_TRACE_LOGGER = VertexiumLoggerFactory.getLogger(SCROLL_API_STACK_TRACE_LOGGER_NAME);
     private final Long limit;
     private QueryResultsIterable<T> firstIterable;
     private boolean initCalled;
@@ -109,6 +116,8 @@ public abstract class InfiniteScrollIterable<T> implements QueryResultsIterable<
 
     private class InfiniteIterator implements CloseableIterator<T> {
         private final String scrollId;
+        private final StackTraceElement[] stackTrace;
+        private boolean scrollIdClosed;
         private Iterator<T> it;
         private T next;
         private T current;
@@ -117,6 +126,11 @@ public abstract class InfiniteScrollIterable<T> implements QueryResultsIterable<
         public InfiniteIterator(String scrollId, Iterator<T> it) {
             this.scrollId = scrollId;
             this.it = it;
+            if (SCROLL_API_STACK_TRACE_LOGGER.isTraceEnabled()) {
+                stackTrace = Thread.currentThread().getStackTrace();
+            } else {
+                stackTrace = null;
+            }
         }
 
         @Override
@@ -173,8 +187,31 @@ public abstract class InfiniteScrollIterable<T> implements QueryResultsIterable<
         @Override
         public void close() {
             closeScroll(scrollId);
+            scrollIdClosed = true;
             CloseableUtils.closeQuietly(it);
             InfiniteScrollIterable.this.close();
+        }
+
+        @Override
+        protected void finalize() throws Throwable {
+            if (!scrollIdClosed) {
+                LOGGER.warn(
+                    "Elasticsearch scroll not closed. This can occur if you do not iterate completly or did not call close on the iterable.%s",
+                    stackTrace != null && SCROLL_API_STACK_TRACE_LOGGER.isTraceEnabled()
+                        ? ""
+                        : String.format(" To enable stack traces enable trace logging on \"%s\"", SCROLL_API_STACK_TRACE_LOGGER_NAME)
+                );
+                if (stackTrace != null) {
+                    SCROLL_API_STACK_TRACE_LOGGER.trace(
+                        "Source of unclosed iterable:\n  %s",
+                        Arrays.stream(stackTrace)
+                            .map(StackTraceElement::toString)
+                            .collect(Collectors.joining("\n  "))
+                    );
+                }
+                close();
+            }
+            super.finalize();
         }
     }
 }